+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Updated the requests variable from a object to an array.
 
 ## v0.0.8
 - Model post requests.