+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Added the body parameters to the post request.
+- Updated the requests variable from a object to an array.
 
 ## v0.0.8
 - Model post requests.