+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Added the body parameters to the post request.
 - Updated the requests variable from a object to an array.
 
 ## v0.0.8