+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 - Added option to override the config of an request.
+- Added the body parameters to the post request.
+- Updated the requests variable from a object to an array.
 
 ## v0.0.8
 - Model post requests.